shinyDashboardLogoDIY {dashboardthemes}	R Documentation

Creates a custom logo object for a shinydashboard application

Description

Creates a custom logo object for a shinydashboard application

Usage

shinyDashboardLogoDIY(boldText, mainText, textSize = 15, badgeText,
  badgeTextColor, badgeTextSize = 2, badgeBackColor,
  badgeBorderRadius = 3)

Arguments

boldText
String. Bold text for the logo.

mainText
String. Main text for the logo.

textSize
Numeric. Text size for the logo. Defaults to 15.

badgeText
String. Text for the logo badge.

badgeTextColor
String. Text color of the logo badge.

badgeTextSize
Numeric. Text color of the logo badge. Defaults to 2.

badgeBackColor
String. Background color of the logo badge.

badgeBorderRadius
Numeric. Border radius of the logo badge. Defaults to 3.

Value

HTML code. Logo for shinydashboard's sidebar.

Note

This logo design was inspired from http://www.dataseries.org/.

See Also

shinyDashboardLogo, cssGradientThreeColors

Examples

logo_blue_gradient <- shinyDashboardLogoDIY(
  boldText = "SD"
  ,mainText = "Themes"
  ,textSize = 16
  ,badgeText = "BETA"
  ,badgeTextColor = "white"
  ,badgeTextSize = 2
  ,badgeBackColor = "#40E0D0"
  ,badgeBorderRadius = 3
)

[Package dashboardthemes version 1.0.5 Index]